api/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
python-jose
passlib
sqlalchemy
decouple
